% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minESS.R
\name{minESS}
\alias{minESS}
\title{Minimum Effective Sample Size}
\usage{
minESS(p, alpha = 0.05, epsilon = 0.05)
}
\arguments{
\item{p}{Number of parameters being sampled, e.g. (beta1, beta2, beta3) implies p = 3}

\item{alpha}{Significance level used to compute ESS; defaults to alpha = 0.05 i.e. 5\%}

\item{epsilon}{Relative precision term; fixing all other elements, as precision is set smaller, sample size increases; defaults to 0.05}
}
\description{
This function calculates minimum effective sample size for a MCMC sample.
This is adapted from the more complex version in stableGR:::minESS.
}
\examples{
minESS(p = 2, alpha = 0.05, epsilon = 0.05)
minESS(p = 2, alpha = 0.10, epsilon = 0.05)
}
\references{
D. Vats, J. Flegal, & G. Jones 2019. Multivariate output analysis for Markov chain Monte Carlo.
Biometrika, Volume 106, Issue 2, June 2019,
Pages 321–337, https://doi.org/10.1093/biomet/asz002

D. Vats and C. Knudson.
Revisiting the Gelman-Rubin Diagnostic. https://arxiv.org/abs/1812.09384
}
