% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_psrf.R
\name{target_psrf}
\alias{target_psrf}
\title{Target potential scale reduction factor (PSRF)}
\usage{
target_psrf(m, p, alpha = 0.05, epsilon = 0.05)
}
\arguments{
\item{m}{Number of MCMC chains, e.g. 3 chains implies m = 3}

\item{p}{Number of parameters being sampled, e.g. (beta1, beta2, beta3) implies p = 3}

\item{alpha}{Significance level used to compute ESS; defaults to alpha = 0.05 i.e. 5\%}

\item{epsilon}{Relative precision term; fixing all other elements,
as precision is set smaller, sample size increases; defaults to 0.05}
}
\description{
This function calculates the target PSRF for a set of MCMC chains.
This is adapted from the more complex version in stableGR
}
\examples{
target_psrf(m = 2, p = 2, alpha = 0.05, epsilon = 0.05)
target_psrf(m = 5, p = 2, alpha = 0.10, epsilon = 0.05)
}
\references{
D. Vats and C. Knudson. Revisiting the Gelman-Rubin Diagnostic.
https://arxiv.org/abs/1812.09384
}
